+++ NEWS.md
@@ -2,6 +2,7 @@
 # remaCor 0.0.11
  - Oct 11, 2022
  - Handle errors when covariance matrix is not invertable
+ - center title for `plotForest()`
 
 # remaCor 0.0.10
  - Sept 7, 2022